Move notification checks out of ExecutablePlan
As these are now done on the logical plan they are no longer
a concern of the runtime, but can be done directly after planning.
