+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/compatibility/v3_4/runtime/executionplan/ExecutionPlan.scala
@@ -21,9 +21,8 @@ package org.neo4j.cypher.internal.compatibility.v3_4.runtime.executionplan
 
 import org.neo4j.cypher.internal.compatibility.v3_4.runtime.RuntimeName
 import org.neo4j.cypher.internal.frontend.v3_4.PlannerName
-import org.neo4j.cypher.internal.frontend.v3_4.notification.InternalNotification
 import org.neo4j.cypher.internal.frontend.v3_4.phases.CacheCheckResult
-import org.neo4j.cypher.internal.planner.v3_4.spi.{GraphStatistics, PlanContext}
+import org.neo4j.cypher.internal.planner.v3_4.spi.GraphStatistics
 import org.neo4j.cypher.internal.runtime.{ExecutionMode, InternalExecutionResult, QueryContext}
 import org.neo4j.cypher.internal.v3_4.logical.plans.IndexUsage
 import org.neo4j.values.virtual.MapValue
@@ -34,6 +33,5 @@ abstract class ExecutionPlan {
   def plannerUsed: PlannerName
   def isStale(lastTxId: () => Long, statistics: GraphStatistics): CacheCheckResult
   def runtimeUsed: RuntimeName
-  def notifications(planContext: PlanContext): Seq[InternalNotification]
   def plannedIndexUsage: Seq[IndexUsage] = Seq.empty
 }
